refactor: 유틸리티 패키지 이동
