+++ src/test/java/roomescape/config/TestClockConfig.java
@@ -4,7 +4,7 @@ import java.time.Clock;
 import java.time.LocalDateTime;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import roomescape.domain.reservation.FixedClock;
+import roomescape.domain.reservation.utils.FixedClock;
 
 @TestConfiguration
 public class TestClockConfig {
