+++ src/test/java/roomescape/domain/reservation/ThemeApiTest.java
@@ -22,7 +22,7 @@ import roomescape.domain.reservation.entity.Theme;
 import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.reservation.repository.ReservationTimeRepository;
 import roomescape.domain.reservation.repository.ThemeRepository;
-
+import roomescape.domain.reservation.utils.JdbcTemplateUtils;
 
 @SpringBootTest(webEnvironment = WebEnvironment.DEFINED_PORT)
 class ThemeApiTest {
