+++ src/test/java/roomescape/domain/reservation/repository/impl/ReservationTimeDAOTest.java
@@ -20,7 +20,7 @@ import roomescape.common.exception.EntityNotFoundException;
 import roomescape.config.TestConfig;
 import roomescape.domain.reservation.entity.ReservationTime;
 import roomescape.domain.reservation.repository.ReservationTimeRepository;
-import roomescape.domain.reservation.JdbcTemplateUtils;
+import roomescape.domain.reservation.utils.JdbcTemplateUtils;
 
 @Disabled
 class ReservationTimeDAOTest {
